lmodrc.lua: fasrc's custom scDescriptT (FIXME)
Not sure if we really need this but adding just for completeness.
Disocvered this while comparing files in init from our manually edited
install vs lmod 5.4.1 git tag. This is ugly given that we're hardcoding
paths here so at the very least need to convert this to read from
environment (perhaps using $FASRCSW_PROD?).